+++ src/stock-physical-inventory-draft/physical-inventory-draft.module.js
@@ -33,6 +33,7 @@
         'stock-constants',
         'stock-reasons-modal',
         'openlmis-auth',
-        'stock-physical-inventory'
+        'stock-physical-inventory',
+        'stock-card'
     ]);
 })();